MyFlaskApp was added
