+++ MyFlaskApp/templates/about.html
@@ -1,0 +1,9 @@
+{% extends 'layout.html' %}
+
+{% block body %}
+    <div class="jumbotron text-center">
+        <h1>About this..</h1>
+        <h2>It's simple, type and post commnet for fun </h2>
+    </div>
+
+{% endblock %}
