+++ MyFlaskApp/templates/includes/_navbar.html
@@ -1,0 +1,16 @@
+
+<!-- My navbar for the header -->
+
+<nav class="navbar navbar-default">
+    <div class="container">
+        <div class="navbar-header">
+            <a class="navbar-brand">Umber's FlaskApp</a>
+        </div>
+        <div id="navbar" class="collapse navbar-collapse">
+            <ul class="nav navbar-nav">
+                <li><a href="/">Index</a></li>
+                <li><a href="/about">About</a></li>
+            </ul>
+        </div><!-- /.nav-collapse -->
+    </div>
+</nav>
