+++ MyFlaskApp/templates/index.html
@@ -1,0 +1,26 @@
+{% extends 'layout.html' %}
+
+{% block body %}
+
+    <div class="jumbotron text-center">
+        <h1>Welcome to Umber's FlaskApp</h1>
+        <p class="lead">This application is built on the Python Flask framework, Enjoy!..</p>
+        <p>Don't be rude, be polite, and have fun. </p><br/><br/>
+
+        <div class="container text-left">
+        {% for comment in comments %}
+                <div class="row">
+                    {{ comment }}
+                </div>
+            {% endfor %}
+
+            <div class="row">
+                <form action="." method="POST">
+                    <textarea class="form-control" name="contents" placeholder="Enter a comment"></textarea>
+                    <input type="submit" class="btn btn-success" value="Post comment">
+                </form>
+            </div>
+        </div>
+    </div>
+
+{% endblock %}
