+++ MyFlaskApp/templates/layout.html
@@ -1,0 +1,25 @@
+<html>
+    <head>
+        <meta charset="utf-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+
+<!--Bootstrap-->
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css"
+        integrity="sha512-dTfge/zgoMYpP7QbHy4gWMEGsbsdZeCXz7irItjcC3sPUFtf0kuFbDz/ixG7ArTxmDjLXDmezHubeNikyKGVyQ=="
+        crossorigin="anonymous">
+<!--Bootstrap-->
+
+        <title>MyFlaskApp</title>
+    </head>
+
+    <body>
+        {% include 'includes/_navbar.html' %}
+
+        {% block body %}
+
+
+        {% endblock %}
+
+    </body>
+</html>
