+++ homework/TorresHumbertoInitial3.py
